maintain: move Txn types to query.go
And remove the TableName method.
